Utilize drawSvg from canvg
It runs quickly enough that I don't think anyone will notice. It does,
unfortunately, not use the already downloaded SVG image. Also, some of
the SVG images have a shifted viewBox, messing up the alignment. I
already knew about that, but neglected to think about it in terms of
displaying the SVG.